requirements.txt
django==2.1.5
tinymce4-lite==1.8.0
